add pagination category

+++ shop/persistence/src/main/java/com/fursel/persistence/repository/TenantRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fursel.persistence.Tenant;
 
-public interface TenantRepository extends JpaRepository<Tenant, Integer> {
+public interface TenantRepository extends JpaRepository<Tenant, Long> {
 
 }
