+++ shop/persistence/src/main/java/com/fursel/persistence/repository/TenantUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.fursel.persistence.TenantUser;
 
-public interface TenantUserRepository extends JpaRepository<TenantUser, Integer> {
+public interface TenantUserRepository extends JpaRepository<TenantUser, Long> {
 
 	@Query("select c from TenantUser c where c.email = :email")
 	public TenantUser findByEmail(@Param("email") String email);
